refactor(api): déplacer les services de commentaires d'action vers actionComment/ refers to https://trello.com/c/evsOelCU/2717
Les services createComment, updateComment, deleteComment, getCommentReport
et enrichCommentsAttachments vivaient mélangés dans services/action/, aux
côtés du CRUD des actions elles-mêmes, alors que le domaine shantytown
dispose déjà de sa propre séparation (services/shantytownComment/). Création
d'un barrel actionComment/index.ts symétrique et mise à jour de tous les
imports (actionService.ts, fetch.ts, write.fetchAction.ts et leurs specs,
controllers/actionComment/*).
