+++ packages/api/server/services/action/fetch.ts
@@ -1,8 +1,8 @@
 import actionModel from '#server/models/actionModel/index';
 import ServiceError from '#server/errors/ServiceError';
+import enrichCommentsAttachments from '#server/services/actionComment/enrichCommentsAttachments';
 import { EnrichedAction } from '#root/types/resources/ActionEnriched.d';
 import { User } from '#root/types/resources/User.d';
-import enrichCommentsAttachments from './enrichCommentsAttachments';
 
 export default async function fetch(user: User, actionIds?: number[], organizationId?: number | null): Promise<EnrichedAction[]> {
     const actions = await actionModel.fetch(user, actionIds);
