+++ packages/api/server/services/actionComment/index.ts
@@ -1,0 +1,11 @@
+import createComment from './createComment';
+import updateComment from './updateComment';
+import deleteComment from './deleteComment';
+import getCommentReport from './getCommentReport';
+
+export default {
+    createComment,
+    updateComment,
+    deleteComment,
+    getCommentReport,
+};
